Refactor Story Signup and Submission Logic
- Updated `signUp` function to utilize `updateStorytellerByToken` for cleaner code.
- Simplified `StorySignupPage` by removing unnecessary state management and form handling, delegating to `StorySignupForm` component.
- Enhanced `StorySubmitForm` to use `upsert` for story saving, streamlining the logic for existing and new stories.
- Removed redundant manual updates to storyteller records after story submission, relying on database triggers instead.
- Improved footer generation in PDF reports for consistent page numbering and branding.
